+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `PauliNet`:
+    - `mo_factory`, functionality replaced with a mean-field backflow
+
 ## [0.2.0] - 2020-08-19
 
 ### Added